19/06/11:04
Version testable et stable !
Les barres de vie sont bien reliées, les dégâts s'infligent bien, pensez juste à ne pas right-click un ennemi ! Pleins d'erreurs en découleront.

+++ Assets/Scripts/ClickOnMe.cs
@@ -14,7 +14,7 @@ public class ClickOnMe : MonoBehaviour
     public void takeDamage(Charac attack , Charac target)
     {
 
-        int mult;
+        float mult;
         if (target.armor.GetValue() > 0) mult = 100 / (100 + target.armor.GetValue());
         else mult = 2 - 100 / (100 - target.armor.GetValue());
         target.currentHp -= attack.damage.GetValue() * mult;
